Ajout Movieet supression Dog

+++ src/main/java/com/repositories/MovieRepository.java
@@ -4,6 +4,7 @@ import com.entities.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+// On remplace Long par Integer ici
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Long> {
+public interface MovieRepository extends JpaRepository<Movie, Integer> {
 }
